+++ CHANGELOG.md
@@ -23,7 +23,8 @@
     - Removed extended gamma levels and `gammacycle`
     - Changed `menu_background_darkening` to `menu_backdrop_darkening` [2]
     - Rearranged menus
-  - **[Woof! 14.0.1](https://github.com/fabiangreffrath/woof/releases/tag/woof_14.0.1)**
+  - **[Woof! 14.0.1](https://github.com/fabiangreffrath/woof/releases/tag/woof_14.0.1)**, note:
+    - Maintained SDL render driver setting (`sdl_renderdriver`)
 - **Reimplemented _Move Over/Under Things_ feature** [3], making it much less bug-prone
 - **Renamed _Impact Pitch_ (`impact_pitch`) to _Flinching_ (`flinching`)** [2]
 - **Tweaked zooming effect**